Move wasm_import_module_map provider to cg_ssa
